+++ src/Models/ModelRefreshToken.php
@@ -9,14 +9,16 @@ use Illuminate\Database\Eloquent\Model;
  */
 class ModelRefreshToken extends Model
 {
-    protected $fillable = ['model_id', 'model_type', 'refresh_token', 'expires_at', 'blacklisted'];
+    protected $fillable = ['model_id', 'model_type', 'device_id', 'refresh_token', 'expires_at', 'blacklisted'];
 
-    public function getRefreshToken($modelType, $modelId)
+    public function getRefreshToken($modelType, $modelId, $deviceId = null)
     {
         return ModelRefreshToken::where(function ($query) use ($modelId) {
             $query->where('model_id', '=', $modelId);
         })->where(function ($query) use ($modelType) {
             $query->where('model_type', '=', $modelType);
+        })->where(function ($query) use ($deviceId) {
+            $query->where('device_id', '=', $deviceId);
         })
         ->orderBy('expires_at', 'desc')
         ->first();
@@ -27,7 +29,8 @@ class ModelRefreshToken extends Model
         return ModelRefreshToken::updateOrCreate(
             [
                 'model_id' => $refreshTokenData['model_id'], 
-                'model_type' => $refreshTokenData['model_type']
+                'model_type' => $refreshTokenData['model_type'],
+                'device_id' => $refreshTokenData['device_id']
             ],
             [
                 'refresh_token' => $refreshTokenData['refresh_token'],